Add Retry module for rescheduling failed builds on a different distribution
The primary use case for this is regression testing, where we want
to see if a set of packages causes build failures that do not happen
without them.

Having two distributions that differ only in these packages, we
can automatically trigger a retry on the 'base' distribution if
there's a failure on the 'modified' one.

The module is trivial otherwise, but there's an unfortunate race condition
where when we requeue a build from a post build hook, another thread
starts sbuild but the original one proceeds to clean up its files
including the .dsc in the incoming directory. This results in sbuild
failing in a very confusing way.

This is currently worked around by inserting a sleep call after requeuing
so the new sbuild can get its act together before the hooks are finished.
A better fix would probably be to have a distribution specific place
where sbuild downloads its files.