Pass --no-arch-all to sbuild when building a binNMU
This is a workaround for https://bugs.debian.org/898946 in sbuild.

Debian binNMUs are not currently defined for Architecture:all packages,
but sbuild currently builds them by default unless --no-arch-all is
passed.

The resulting +b1_all.deb binaries can confuse apt so that it considers
the arch-dep packages from the same source uninstallable when they depend
on the non-binNMU-versioned arch-indep packages.